Remove separate payments branch
